fix(gdn): carry the bf16-rounded projection as the next column's convolution tap
The fused convolution epilogue publishes each column's projection output rounded to bf16 -- into
the recurrent state on the decode path, into the replay record on the verify path -- and then
carries the UNROUNDED fp32 value forward as the in-window tap for the following column. A
sequential decode never sees that value: it reads its taps back from bf16 state. So a multi-column
verify was not computing the same number a different way, it was computing a different number, at
bf16 magnitude on thirty of forty layers, on every column but the first.

That asymmetry matched the symptom. Column 0's taps all come from state and are clean, so a
rejected round -- which commits argmax of column 0 -- saw only fp32-level noise, and structured
output stayed byte-identical to non-speculative decode at every draft width. Columns 1 and up
carried the error, which is exactly the accepted draft and the bonus token, and free prose diverged
from k=1 onward.

The rounded value is already computed one line above, so the fix is a reuse, not a new conversion.
T=1 is untouched by construction: the store template instantiates with Tokens=1 on the decode path,
where the loop runs once and the tap assignment is dead.

Measured, prose at k=3, three runs per arm, spread under 0.1%:
  with    513.96 / 513.70 / 513.90 tok/s, acceptance 2.49
  without 509.82 / 509.61 / 509.52 tok/s, acceptance 2.47
so +0.8% on both -- a more faithful verify agrees with the drafter slightly more often.

This does NOT make speculative decode output-identical to sequential decode on prose. The dense
projections and the MoE router still reduce differently at T=1 and T>=2 (SIMT fp32 lane chains
against tensor-core mma, and an 8-warp against a 4-warp router reduction), and those are fp32
reassociations that no partition choice reconciles. Fixing the one genuine precision difference
does not remove them; it removes the only one that was thirty thousand times larger.

(cherry picked from commit 0a87c1926453d5f53138dea469c252079a271c02)

+++ src/ops/gdn_input_proj/gdn_conv.cuh
@@ -111,9 +111,14 @@ struct GdnConvEpilogue {
             }
 
             publish.publish(token, batch_row, row, s1, s2, p);
+            // The tap this column leaves for the next one has to be the value the next
+            // column would have READ, and a sequential decode reads it back from bf16 state.
+            // Carrying the unrounded projection forward instead makes a multi-column pass
+            // compute a different number than one column at a time -- bf16-magnitude, on
+            // thirty of forty layers, on every column but the first.
             s0 = s1;
             s1 = s2;
-            s2 = p;
+            s2 = __bfloat162float(__float2bfloat16_rn(p));
         }
     }
 };
